move structs to header
